Update MainPage.xaml.cs
In this file you will see the main page / main base of the app because we have added navigation page to signing page which will direct you to sign in page formality and then if you are successful in signing then it will again direct you to this page for main page of start learning buttons and then you will see the dashboard which is the home page of this page. for now we have added this things only in this app.