LearningPage
A ContentPage element was defined with the namespace declarations for Xamarin Maui (http://schemas.microsoft.com/dotnet/2021/maui) and the XAML namespace (http://schemas.microsoft.com/winfx/2009/xaml).
The ContentPage class was assigned the LearnEase.LearningPage class using the x:Class attribute.
A ToolbarItem was added to the ToolbarItems collection of the ContentPage to create a toolbar button labeled "More". The Clicked event handler for this button is set to OnMoreClicked.
Within the ContentPage.Content section, a StackLayout was added to contain the page content.
Inside the StackLayout, a Label element was added to display the text "Welcome to the Learning Page!". Various properties such as HorizontalOptions, VerticalOptions, FontSize, and Margin were set to customize the appearance of the label.
A ScrollView was added around the StackLayout to enable scrolling functionality.
A Image element was added to display an image. The Source property was set to the path of the image file added to the project. Additionally, properties like Aspect, HeightRequest, and WidthRequest were set to customize the appearance of the image.