LearningPage
Event Handlers: If there's an event handler defined for the "More" button's Clicked event, it would be implemented in this file. For example, the OnMoreClicked method mentioned in the XAML file would be implemented here.
Constructor: The constructor of the LearningPage class might contain initialization logic or additional setup for the page.
Methods: Any other methods specific to the functionality of the page could be defined here.
Properties: Properties related to the page could be declared here.
Data Binding: If there's any data binding between the UI elements and the underlying data model, it would be set up in this file.
Navigation: Navigation-related logic, such as navigating to other pages, could be implemented here.